Complete ICML paper with real experiment results and figures
All 5 experiments completed with LoRA learning verified (B-norms 0.43-0.73):
- uniform_r4: DINO=0.371, CLIP-T=0.271, 797K params
- baseline_r8: DINO=0.398, CLIP-T=0.265, 1.59M params
- uniform_r16: DINO=0.378, CLIP-T=0.265, 3.19M params
- blockwise_no_ccd: DINO=0.393, CLIP-T=0.271, 1.49M params (53% fewer than r16)
- blockwise_ccd: DINO=0.397, CLIP-T=0.267, 1.49M params

Paper includes 4 tables with verified real data, 2 figures, full
bibliography with verified citations. All claims cross-checked against
actual experiment outputs.

Bug fixes in this commit:
- trainer.py: Cast CCD features to float32 for mixed-precision compat
- blockwise_lora.py: Mixed-precision forward (fp16 input + fp32 LoRA)

+++ modularbooth/training/trainer.py
@@ -906,8 +906,11 @@ class ModularBoothTrainer:
         elif hasattr(ccd_loss_module, "ccd_loss"):
             ccd_loss_module = ccd_loss_module.ccd_loss
 
+        # Ensure all features share the same dtype (fp32 LoRA can produce mixed).
         ccd_loss = ccd_loss_module(
-            subject_features, positive_features, negative_features
+            subject_features.float(),
+            positive_features.float(),
+            negative_features.float(),
         )
         return ccd_loss
